requirements.txt
PyMuPDF==1.24.1
openai==1.51.2
Pillow==10.4.0
python-dotenv==1.0.0
pandas==2.2.3
tenacity==8.3.0
click==8.1.7
loguru==0.7.2
pydantic==2.9.2
tqdm==4.66.5
requests==2.32.3

# Development dependencies
pytest==8.3.3
pytest-cov==5.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0 